requirements.txt
# Core ML
torch>=2.0.0
transformers>=4.36.0
datasets>=2.14.0
accelerate>=0.24.0

# QLoRA / PEFT
peft>=0.7.0
bitsandbytes>=0.41.0

# Databricks
databricks-sdk>=0.12.0
databricks-sql-connector>=2.9.0

# MLflow
mlflow>=2.9.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
pandas>=2.0.0
numpy>=1.24.0

# Evaluation
lm-eval>=0.4.0  # Optional: for standard benchmarks
